+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ Development
 ======================
 * Added logging parameter in :func:`dynts.evaluate`.
 * Formatters are instances rather than functions.
+* Can specify backend in :func:`dynts.evaluate`.
+* Introduced :func:`dynts.tsname` for creating names for a mutivariate timeseries.
+* Introduced :func:`dynts.merge` for merging two or more :class:`dynts.TimeSeries`.
+* **60 tests**. 
 
 Version 0.2 - 2010 September 24
 ====================================
@@ -11,7 +15,6 @@ Version 0.2 - 2010 September 24
 * Added ccy_ to dependencies.
 * **4** dsl functions.
 * **53 tests**.
-
 
 Version 0.1.a2  - 2010 September 12
 ====================================